fix(ui): localization for constant files
