+++ openmetadata-ui/src/main/resources/ui/src/components/app-bar/Appbar.tsx
@@ -35,7 +35,7 @@ import {
   urlGitbookDocs,
   urlGithubRepo,
   urlJoinSlack,
-} from '../../constants/url.const';
+} from '../../constants/URL.const';
 import { useAuth } from '../../hooks/authHooks';
 import jsonData from '../../jsons/en';
 import {
