+++ openmetadata-ui/src/main/resources/ui/src/pages/PoliciesPage/PoliciesDetailPage/EditRulePage.tsx
@@ -19,7 +19,7 @@ import { useHistory, useParams } from 'react-router-dom';
 import { getPolicyByName, patchPolicy } from '../../../axiosAPIs/rolesAPIV1';
 import TitleBreadcrumb from '../../../components/common/title-breadcrumb/title-breadcrumb.component';
 import Loader from '../../../components/Loader/Loader';
-import { GlobalSettingOptions } from '../../../constants/globalSettings.constants';
+import { GlobalSettingOptions } from '../../../constants/GlobalSettings.constants';
 import { Effect, Rule } from '../../../generated/api/policies/createPolicy';
 import { Policy } from '../../../generated/entity/policies/policy';
 import { getEntityName } from '../../../utils/CommonUtils';
