+++ openmetadata-ui/src/main/resources/ui/src/pages/database-details/index.tsx
@@ -70,7 +70,7 @@ import {
   pagingObject,
 } from '../../constants/constants';
 import { EntityField } from '../../constants/Feeds.constants';
-import { GlobalSettingsMenuCategory } from '../../constants/globalSettings.constants';
+import { GlobalSettingsMenuCategory } from '../../constants/GlobalSettings.constants';
 import { observerOptions } from '../../constants/Mydata.constants';
 import { EntityType, TabSpecificField } from '../../enums/entity.enum';
 import { ServiceCategory } from '../../enums/service.enum';
